Fixed: exception thrown when adding empty group

+++ lib/watcher_groups_issue_patch.rb
@@ -27,8 +27,8 @@ module WatcherGroupsIssuePatch
 
     def watcher_groups_ids=(group_ids)
       groups = group_ids.collect {|group_id| Group.find(group_id) if Group.find(group_id).is_a?(Group)  }
-      user_ids = groups.map(&:users).flatten.compact.uniq.map(&:id)
-      Watcher.where("watchable_type = '#{self.class}' AND watchable_id = #{self.id} AND user_id IN (#{user_ids.join(',')})").delete_all
+			user_ids = groups.map(&:users).flatten.compact.uniq.map(&:id)
+      Watcher.where("watchable_type = '#{self.class}' AND watchable_id = #{self.id} AND user_id IN (#{user_ids.join(',')})").delete_all unless user_ids.empty?
       groups.each do |group|
         self.add_watcher_group(group)
       end
